[fix] Do *not* log out on props change
The `useAsyncEffect` second parameter (the dependency list) was clearly too broad, as it contained an object that gets spuriously reconstructed even when none of its fields change (this is a symptom of the problem described in https://react.dev/reference/react/useEffect#removing-unnecessary-object-dependencies ). This resulted in a bug whence any change in any unrelated prop (including `children`) would cause a logout.

Upon further reflection on the desirablity (or rather, lack thereof) of the logout behavior, make it so `OIDCContext` avoids doing that inasmuch as possible; that is, set the `useAsyncEffect` dependency list to empty.